refactor: extract PBL Task Page mobile components and utilities
BREAKING CHANGES RESOLVED: Reduced page.tsx from 854 to 394 lines (54% reduction)

Extracted Components:
- MobileProgressView (progress tracking sidebar)
- MobileTaskInfoView (task details and evaluation)
- MobileChatView (chat interface with AI)
- MobileBottomNavigation (mobile navigation tabs)

Extracted Utilities (message-handlers.ts):
- createProgramIfNeeded: Handle program creation
- saveInteraction: Save user/AI interactions
- getAIResponse: Fetch AI chat responses
- navigateToNextTask: Task navigation logic

Benefits:
- Single Responsibility: Each component has one clear purpose
- Reusability: Mobile components can be used elsewhere
- Maintainability: Easier to test and modify
- Type Safety: Proper TypeScript types throughout

Files Modified:
- src/app/pbl/scenarios/[id]/programs/[programId]/tasks/[taskId]/page.tsx (394 lines, was 854)

Files Created:
- src/components/pbl/task/mobile/MobileProgressView.tsx
- src/components/pbl/task/mobile/MobileTaskInfoView.tsx
- src/components/pbl/task/mobile/MobileChatView.tsx
- src/components/pbl/task/mobile/MobileBottomNavigation.tsx
- src/app/pbl/scenarios/[id]/programs/[programId]/tasks/[taskId]/utils/message-handlers.ts
